requirements.txt
Flask>=2.0.0
Flask-CORS>=4.0.0
requests>=2.28.0
beautifulsoup4>=4.11.0
pandas>=1.5.0
python-dateutil>=2.8.0
pycountry>=22.3.5
lxml>=4.9.0
openpyxl>=3.1.0

soupsieve~=2.8.3
typing_extensions~=4.14.1
pip~=24.0
MarkupSafe~=3.0.2
Werkzeug~=3.1.3
click~=8.2.1
Jinja2~=3.1.6
blinker~=1.9.0
itsdangerous~=2.2.0
numpy~=2.4.2
charset-normalizer~=3.4.4
certifi~=2026.1.4
tzlocal~=5.3.1
urllib3~=2.6.3
six~=1.17.0
et_xmlfile~=2.0.0
idna~=3.11
APScheduler~=3.11.2